requirements.txt
mysql-connector-python
tk
matplotlib
scikit-learn
#pip install -r requirements.txt
